requirements.txt
flask
flask_session
sqlalchemy
python-dotenv

